+++ moobae/src/main/java/com/example/moobae/domain/mapper/MemberMapper.java
@@ -1,8 +1,8 @@
 package com.example.moobae.domain.mapper;
 
 import com.example.moobae.domain.member.Member;
-import com.example.moobae.domain.member.MemberDTO;
-import com.example.moobae.domain.member.MemberVO;
+import com.example.moobae.domain.member.dto.MemberDTO;
+import com.example.moobae.domain.member.vo.MemberVO;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
