+++ moobae/src/main/java/com/example/moobae/domain/member/vo/MemberVO.java
@@ -1,4 +1,4 @@
-package com.example.moobae.domain.member;
+package com.example.moobae.domain.member.vo;
 
 import lombok.Getter;
